Typo fix: Renamed "_pr_texutre_compute_miplevel" to "_pr_texture_compute_miplevel".

+++ test/win32/main.c
@@ -491,7 +491,7 @@ int main()
             #endif
 
             // Draw floor
-            int floorSize = 100;
+            float floorSize = 100.0f;
             int floorTC = 50;
 
             prLoadIdentity(worldMatrix);
